Move last `web-admin/components` to `features`
