+++ web-admin/src/features/errors/error-store.ts
