+++ web-admin/src/features/navigation/BreadcrumbItem.svelte
