+++ web-admin/src/features/navigation/OrganizationAvatar.svelte
