+++ web-admin/src/routes/+error.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { page } from "$app/stores";
-  import { createErrorPagePropsFromRoutingError } from "../components/errors/error-utils";
-  import ErrorPage from "../components/errors/ErrorPage.svelte";
+  import { createErrorPagePropsFromRoutingError } from "../features/errors/error-utils";
+  import ErrorPage from "../features/errors/ErrorPage.svelte";
 
   $: errorPageProps = createErrorPagePropsFromRoutingError($page.status);
 </script>
