+++ web-admin/src/routes/+page.svelte
@@ -4,8 +4,8 @@
     createAdminServiceGetCurrentUser,
     createAdminServiceListOrganizations,
   } from "../client";
-  import AuthRedirect from "../components/authentication/AuthRedirect.svelte";
-  import WelcomeMessage from "../components/home/WelcomeMessage.svelte";
+  import AuthRedirect from "../features/authentication/AuthRedirect.svelte";
+  import WelcomeMessage from "../features/home/WelcomeMessage.svelte";
   import OrganizationList from "../features/organizations/OrganizationList.svelte";
 
   const user = createAdminServiceGetCurrentUser();
